refactor export_layer
